Enable Ubuntu 18.04 host support (#974)
* Enable Ubuntu 18.04 and 16.04 host support

Fix some packages that have changed since 16.04.
Remove some supurious checks in checkenv
Fix auth issues in multistrap

Builds end to end on an 18.04 system and should workon on 19.x too

* Enable noauth in multistrap when using PYNQ_UBUNTU_REPO, add support for 18.04 in setup_host

* Fix resolv.conf issue

- Fix resolv.conf issue host-side using setup_host.sh
- Enable usage of insecure repositories in multistrap when PYNQ_UBUNTU_REPO is defined
- Add rsync and python3-pip install and check in setup_host and check_env

* update vagrant file

* fix jupyter lab building

* make ubuntu 16 and 18 setup and check consistent

* ignore dsa files in sdist

+++ sdbuild/scripts/setup_host.sh
@@ -49,13 +49,25 @@ chrpath
 socat
 zlib1g-dev
 zlib1g:i386
-gcc-multilib
 unzip
+rsync
+python3-pip
+python-minimal
+gcc-multilib
 EOT
 set -e
 
 sudo apt-get update
-sudo apt purge -y libgnutls-dev
+
+if [[ $(lsb_release -rs) == "16.04" ]]; then
+    echo "Install packages on Ubuntu 16.04..."
+    sudo apt purge -y libgnutls-dev
+elif [[ $(lsb_release -rs) == "18.04" ]]; then
+    echo "Install packages on Ubuntu 18.04..."
+else
+    echo "Error: current OS not supported."
+    exit 1
+fi
 
 # Setup docker and containerd using repository before installing them
 sudo apt-get install -y \
@@ -117,6 +129,12 @@ cd /usr/bin
 if ! which gmake
 then
   sudo ln -s make gmake
+fi
+
+# make sure resolv.conf is where it's going to be expected in chroot
+if [ ! -f /run/systemd/resolve/stub-resolv.conf ]; then
+    sudo mkdir -p /run/systemd/resolve
+    sudo cp -L /etc/resolv.conf /run/systemd/resolve/stub-resolv.conf
 fi
 
 # update setuptools
